# Conflicts:
#	src/main/java/com/tms/web/services/busines/impl/UploadBookServiceImpl.java
#	src/main/java/com/tms/web/services/populators/Populator.java
#	src/main/resources/application.properties

+++ src/main/java/com/tms/web/services/busines/impl/UploadBookServiceImpl.java
@@ -5,6 +5,9 @@ import com.tms.web.entities.library.Book;
 import com.tms.web.services.busines.BookComplexService;
 import com.tms.web.services.busines.ParseFB2BookService;
 import com.tms.web.services.busines.UploadBookService;
+import com.tms.web.services.validators.Validator;
+import com.tms.web.services.validators.impl.FileValidatorImpl;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -16,43 +19,51 @@ import java.io.IOException;
 import java.util.Objects;
 
 @Service
+@Log4j2
 public class UploadBookServiceImpl implements UploadBookService {
     @Autowired
     private ParseFB2BookService parseFB2BookService;
     @Autowired
     private BookComplexService bookComplexService;
+    @Autowired
+    private FileValidatorImpl fileValidator;
+    @Autowired
+    private Validator<MultipartFile> multipartFileValidator;
 
     @Override
     public Book upload(MultipartFile multipartFile) throws IOException {
         Book book = null;
-        if (multipartFile != null) {
-            String name = multipartFile.getName();
-            System.out.println("File " + name + " has " + multipartFile.getBytes().length + " bytes");
-            long threadId = Thread.currentThread().getId();
-            String threadName = Thread.currentThread().getName();
+        if (multipartFileValidator.validate(multipartFile)){
+            if (multipartFile != null) {
+                log.debug("File " + multipartFile.getName() + " has " + multipartFile.getBytes().length + " bytes");
+                long threadId = Thread.currentThread().getId();
+                String threadName = Thread.currentThread().getName();
 //            ClassLoader classLoader = getClass().getClassLoader();
 //            File bookCreated = new File(classLoader.getResource(".").getFile() + "/bookCreated" + id + ".tmp");
-            File bookCreated = null;
-            try {
-                bookCreated = File.createTempFile("tmpBook", threadName + "_" + String.valueOf(threadId));
-                if (bookCreated.exists()) {
-                    System.out.println("temporary book file created in: " + bookCreated);
-                    multipartFile.transferTo(bookCreated);
-                    book = parseAndSaveBookToDB(bookCreated);
-                    if (bookCreated.delete()) {
-                        System.out.println("temporary book file deleted");
-                    }
-                }
-            } catch (RuntimeException re) {
-                System.out.println("Exception:" + re.getMessage());
-            } finally {
-                if (Objects.nonNull(bookCreated)) {
+                File bookCreated = null;
+                try {
+                    bookCreated = File.createTempFile("tmpBook", threadName + "_" + String.valueOf(threadId));
                     if (bookCreated.exists()) {
-                        bookCreated.delete();
-                        System.out.println("temporary book file deleted");
+                        log.debug("temporary book file created in: " + bookCreated);
+                        multipartFile.transferTo(bookCreated);
+                        book = parseAndSaveBookToDB(bookCreated);
+                        if (bookCreated.delete()) {
+                            log.debug("temporary book file deleted:" + bookCreated);
+                        }
+                    }
+                } catch (RuntimeException re) {
+                    log.fatal("Runtime exception:"+re.getMessage());
+                } finally {
+                    if (Objects.nonNull(bookCreated)) {
+                        if (bookCreated.exists()) {
+                            if (bookCreated.delete()) {
+                                log.debug("temporary book file deleted");
+                            }else {
+                                log.fatal("temporary book file don't deleted");
+                            }
+                        }
                     }
                 }
-            }
 
 
 //            if (bookCreated.createNewFile()) {
@@ -68,14 +79,22 @@ public class UploadBookServiceImpl implements UploadBookService {
 //                System.out.println("Tmp book file deleted");
 //            }
 
+            }
         }
+
         return book;
     }
 
     @Override
     public Book upload(File file) throws IOException {
         Book book = null;
-        book = parseAndSaveBookToDB(file);
+        try {
+            if (fileValidator.validate(file)) {
+                book = parseAndSaveBookToDB(file);
+            }
+        }catch (RuntimeException re){
+            log.fatal("Runtime exception:"+re.getMessage());
+        }
         return book;
     }
 
@@ -86,13 +105,11 @@ public class UploadBookServiceImpl implements UploadBookService {
             book = parseFB2BookService.parse(fb2);
             if (Objects.nonNull(book)) {
                 book = bookComplexService.save(book);
-                System.out.println();
-                if (Objects.isNull(book)) {
-                    throw new IOException("Book non saved");//заменить на свою ошибку!!
-                }
+            }else {
+                log.warn("Fail parse fb2 book to Book.class");
             }
         } catch (ParserConfigurationException | IOException | SAXException e) {
-            throw new IOException("Parsing file not completed");
+            log.fatal(e.getMessage());
         }
         return book;
     }
